requirements.txt
Pillow
PySimpleGUI
tk-tools
numpy
matplotlib
